refactor: parse into `Body` type
For now this just changes the way how HCL is parsed internally. Namely,
HCL text is parsed into a `Body` which itself contains `Attribute` and
`Block` elements.

This is a preparation for enabling HCL parsing without losing context.
The `parse` method and the types in the `structure` module will become
public API eventually.