feat(structure): add `Expression` type (#20)
Adds an `Expression` type to improve the representation of `Attribute`
value expressions.

This is similar to `Value` with the difference that it also allows to
represents expressions that are neither a primitive (e.g. bool) or a
collection (e.g. object) as a `RawExpression`. In the future the
`Expression` enum can be extended to also receive variants for other
expressions like `Conditional` or `FunctionCall`.

For `RawExpression` acts as a catchall variant for those. Its string
representation is an interpolated string (`"${expression}"`).

The `Expression` type will become useful when implementing an HCL
serializer as there is some context needed to serialize expressions
correctly.