refactor: separate serializers and formatting (#51)
With the current implementation where serializers are writing out HCL to the formatter as they go it will get complicated to get the formatting right when the full set of HCL expressions will be added to this crate. I tried multiple approaches but failed a lot.

This change adds serializers for all HCL structure types in the `structure::ser` module. The main serializer in the `ser` module just serializes everything into a `Body` which is then formatted. This allows us to emit better HCL at a later point in time since we have all the context of surrounding structures available when we start formatting. The downside of this is that serialization will be a bit slower due to the additional overhead of serializing into `Body` first, but we might be able to make up for this later by using `Cow` strings.

The `format` module will house the formatting code and eventually the `PrettyFormatter` will be updated and moved from the `ser` module into the `format` module. This will be a breaking change which is out of scope of this PR. The upside of the `format` module is that we'll be able to format the HCL types from this crate more efficiently by simply bypassing serde entirely. Convenience functions for this will be added once the `format` module becomes part of the public API.