feat!: use `VecMap` for object expressions (#72)
BREAKING CHANGE: The underlying map implementation for the `Object<K, V>` type
changed from `IndexMap<K, V>` to `VecMap<K, V>`. For the most common
operations this is a drop-in replacement, but `VecMap` lacks some of the
more exotic APIs the `IndexMap` provides.

This change is necessary to add support for `Expression` as object key
type since it is not `Hash` and not `Ord`.