fix!: prevent creation of infinite and NaN `Number`
BREAKING CHANGE: The `Number` type was changed from an `enum` to an
opaque `struct`. Use `Number::from` to create a number from an integer.
Furthermore, the `From` implementations for `f32` and `f64` were
removed. Use the newly added `Number::from_f64` instead.

This ensures that any number is finite and is almost identical to what
`serde_json` does. Infinite floats get converted to `Value::Null` and
`Expression::Null` respectively when working with the `Value` and
`Expression` types.