feat(format): add `to_string_unchecked` and `strict_mode` (#94)
This is not public API since it might be removed again in favor of the
changes proposed in #93.

For now it's here to ensures that the `to_string` implementations of
`Expression` and `Value` will never ever cause a panic. Since we cannot
signal an error there, it's better to at least do best effort and return
the formatted result even if identifiers are not valid.