chore(format): remove ident validation
Identifiers are validated upon creation and are always valid if created
via `Identifier::new` or `Identifier::sanitized`. The formatter does not
need to validate them again. This also obsoletes the internal strict
mode and `to_string_unchecked` function.

Invalid identifiers can only be created via `Identifier::unchecked`, but
this would be a user error as stated in the method docs.