fix: unclutter crate root exports
Hide some exports at the crate root from the docs. It got pretty crowded
in there and it's really hard to see which types are the most relevant.
Re-exporting every single type from the `expr` and `structure` module
was a mistake. Only the most useful ones should be re-exported.

Furthermore the `block_label!` and `object_key!` macros are hidden now
as well since they do not add much value on their own. They're more
meant for internal usage to build higher level macros like `block!` and
`expression!`.

This is not a breaking change. The exports are still usable but hidden
from the docs with the intention to remove them entirely in a future
release.