fix(eval)!: correctly handle interpolation unwrapping
BREAKING CHANGE: the `Evaluate` implementation of `TemplateExpr` returns
a `Value` instead of a `String` now to support interpolation unwrapping.

Note: this is an empty fix-up commit for #126 which was not detected by
release-please because of an incorrect CCM header. The actual code
change is in 533fb9193f09dc9d274b7ab6d36952f6c17e3394.