feat(format): allow formatting to `String` and `Vec<u8>` directly
This makes using a custom `Formatter` more convenient. The new
`build_vec` method of `FormatterBuilder` returns a `Formatter` with an
internal `Vec<u8>` buffer which can be used in combination with
`Format::format_string` and `Format::format_vec`.

The `Serializer` also received the `serialize_string` and
`serialize_vec` methods for writers implementing `AsMut<Vec<u8>>`.

This change also lowers the number of unsafe usages in this crate from 3
to 1.