chore: convert project into cargo workspace
This change converts the project into a cargo workspace with multiple
crates. The restructuring has a couple of advantages:

- We can release multiple crates with different versions from the same
  repository. This will become relevent for upcoming features like #112
  and #158 which I'd like to separate into purpose-built crates that can
  also be released and used individually.
- Some build tasks can potentially run in parallel.
- Split tooling only used for testing (e.g. `specsuite`/`benchmark`)
  into separate crates to keep things better organized.
- Better control over changes that actually count towards a release.
  Previously, any conventional commit counted towards a release. With
  the restructuring, only changes below `hcl-rs/` (and potentially new
  releaseable crates) will count towards a release, whereas changes in
  below `specsuite/` crate will not. This removes some unnecessary items
  from generated changelogs.

There's probably more.

For the release process to work correctly after the split, we'll need to
switch to release-please's manifest releaser mode, which requires the
configuration to reside in `release-please-config.json` and also needs
`.release-please-manifest.json` to track the current version of each
releaseable crate.

For publishing crates to crates.io, we switch to `cargo-release` as it
can figure out publish ordering for us based on the dependency graph. To
exclude publishing of crates only used for development, we set
`package.publish = false` and `package.metadata.release.release = false`
(relevant for `cargo-release`).