fix(format): do not unescape heredoc strings (#171)
Current logic has `TemplateExpr` doing a `try_unescape` with both a `QuotedString` and `Heredoc` inside  `fmt::Display` This causes side effects when legitimate and valid `\\` characters are contained within a heredoc.

This shows itself with bash scripts, or regular expressions which need escaping. The use case which stumbled upon this was regex in TOML.

```
let content = r###"
config = {
  content = <<-EOT
    www\\.example\\.com
    EOT
}
"###;
println!("{}", hcl::to_string(hcl::from_str(&content)?)?);
```
The output `content` is the string `www\.example\.com` which is incorrect, this PR corrects that.   

As a side note, use of `hcl::eval::from_str` does not exhibit this problem, which is used by the specsuite for loading it's HCL tests, making it incapable of testing this error directly.  Since the `t` file is loaded using `hcl::from_str` which is affected, a test can be done by using a QuotedString for the test, and a Heredoc for the results and vice versa, which is what was done.