perf(parser): avoid `alt` in `array` and `object` parsers (#205)
This improves overall parser performance by around 5%.

Looking for a `for` expression before attempting to parse array/object
items is required due to ambiguities in the grammar.

Until now, this was implemented as a simple `alt` which tries the `for`
parser first. This parser fails in the vast majority of cases because
arrays/objects are way more common than for list/object expressions.

This change replaces the `alt` with a cheaper `peek` that immediately
calls the relevant parser based on the structure of the next 4 bytes
after preceding whitespace instead.