feat(structure)!: prevent duplicate attribute keys in `Body`
BREAKING CHANGE: Various `Body` methods were changed to return
`AttributeMut<'a>`/`StructureMut<'a>` instead of `&'a mut
Attribute`/`&'a mut Structure` to prevent mutable access to attribute
keys. The `VisitMut` trait was updated to reflect these changes as well.

A HCL body must not contain duplicate attribute keys. Multiple instances
of the same block identifier are allowed, though. This change updates
the `Body` type to make it impossible to insert a given attribute key
twice, or to obtain a mutable reference to one.