fix: emit correct empty collection if paths are empty (#446)
There are two kinds of empty collections to emit when paths are empty:

- `[]` if `--file-paths/-P` **is not**  provided
- `{}` if `--file-paths/-P` **is** provided

The current logic didn't account for that and even got it wrong for the
default case, which should emit `[]`.

We'll just remove this fast path since it does not bring any performance
benefit (`bulk_convert` is only called exactly once) but needs to be
kept in sync with the rest of the logic.